Moved around some modules for better hierarchy
