+++ __init__.py
@@ -1,0 +1,5 @@
+
+from .foxpuppet import FoxPuppet
+
+
+__all__ = ['FoxPuppet']
