+++ api/browser/__init__.py
@@ -1,9 +1,3 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
-
-
-from .foxpuppet import FoxPuppet
-
-
-__all__ = ['FoxPuppet']
