+++ foxpuppet.py
@@ -2,8 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from .browser.window import BrowserWindow
-from .windows import Windows
+from .api.browser.window import BrowserWindow
+from .api.windows import Windows
 
 
 class FoxPuppet(object):
