+++ tests/test_browser_model.py
@@ -3,7 +3,7 @@
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
-from foxpuppet import FoxPuppet
+from FoxPuppet.foxpuppet import FoxPuppet
 
 
 class TestBrowserModel(object):
